spacez-visual-intelligence/services/simulator/requirements.txt
nats-py
fastapi
uvicorn
pydantic
websockets
